Refresh file list when doing a description search
