+++ RC-SpeechToText/ClientApp/components/Dashboard/Dashboard.tsx
@@ -7,7 +7,6 @@ import AutomatedFilter from './AutomatedFilter';
 import EditedFilter from './EditedFilter';
 import ReviewedFilter from './ReviewedFilter';
 import MyFilesFilter from './MyFilesFilter';
-import { FileDescriptionSearch } from '../FileView/FileDescriptionSearch';
 import Loading from '../Loading';
 import FilesToReviewFilter from './FilesToReviewFilter';
 
@@ -19,6 +18,7 @@ interface State {
     isEditedFilterActive: boolean,
     isAutomatedFilterActive: boolean,
     isReviewedFilterActive: boolean,
+    searchTerms: string,
     isFilesToReviewFilterActive: boolean,
     loading: boolean,
     unauthorized: boolean
@@ -38,6 +38,7 @@ export default class Dashboard extends React.Component<any, State> {
             isAutomatedFilterActive: false,
             isReviewedFilterActive: false,
             isFilesToReviewFilterActive: false,
+            searchTerms: '',
             loading: false,
             unauthorized: false
         }
@@ -266,6 +267,28 @@ export default class Dashboard extends React.Component<any, State> {
         )
     }
 
+    public searchDescription = () => {
+        const config = {
+            headers: {
+                'Authorization': 'Bearer ' + auth.getAuthToken(),
+                'content-type': 'application/json'
+            }
+        }
+        axios.get('/api/file/getFilesByDescription/' + this.state.searchTerms, config)
+            .then(res => {
+                this.setState({ files: res.data });
+            })
+            .catch(err => {
+                if (err.response.status == 401) {
+                    this.setState({ 'unauthorized': true });
+                }
+            });
+    }
+
+    public handleSearch = (e: any) => {
+        this.setState({ searchTerms: e.target.value })
+    }
+
     public render() {
         return (
             <div className="container">
@@ -315,7 +338,13 @@ export default class Dashboard extends React.Component<any, State> {
                     </div>
 
                     <section className="section column">
-                        <FileDescriptionSearch />
+                        <div>
+                            <div className="field is-horizontal">
+                                <a className="button is-link mg-right-10" onClick={this.searchDescription}> Rechercher </a>
+                                <input className="input" type="text" placeholder="Your search terms" onChange={this.handleSearch} />
+                            </div>
+                        </div>
+
                         <div className="box mg-top-30">
                             {this.state.loading ? <Loading /> : this.state.files ? <FileTable
                                 files={this.state.files}
